kbuild: add CONFIG_CC_OPTIMIZE_FOR_PERFORMANCE_O3
arch/arc/Makefile overrides -O2 with -O3. This is the only user of
ARCH_CFLAGS. There is no user of ARCH_CPPFLAGS or ARCH_AFLAGS.
My plan is to remove ARCH_{CPP,A,C}FLAGS after refactoring the ARC
Makefile.

Currently, ARC has no way to enable -Wmaybe-uninitialized because both
-O3 and -Os disable it. Enabling it will be useful for compile-testing.
This commit allows allmodconfig (, which defaults to -O2) to enable it.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Change-Id: Ifcc76e7e8ad0453e5c48ff7d83bda1c11266432c
Signed-off-by: Jason Edson <jaysonedson@gmail.com>